Dedup products
When a product was bought multiple times (through different bundles), it
will appear multiple times in the JSON data (they only used to appear
once in the pre-API indices). The filename parser will think they are
to be symlinked. Clean up this duplication in the symlink phase by not
producing symlinks for products that would link to themselves.