New homepage and login formats
The new login is somewhat annoying due to the need to extract the form
from a `script` tag of type `text/template`, but it can be worked around
reasonably well with a custom parser that extract the template content
and drops it into the main body before Mechanize parses the result (this
must be done before Mechanize, because Mechanize seems to strip `script`
tags).

The home page change is minor (for our purposes: equal sign instead of
colon).